+++ shop.html
@@ -237,6 +237,6 @@
 
   </footer>
   <script src="https://kit.fontawesome.com/1d67b26f86.js" crossorigin="anonymous"></script>
-  <script src="./assets/JS/index.js" type="module"></script>
+  <script src="./assets/JS/shop.js" type="module"></script>
 </body>
 </html>
